Corrección de las colaboraciones, según lo indicado por Ezequiel.
