+++ src/main/java/domain/donaciones/Distribuir.java
@@ -1,33 +1,30 @@
 package domain.donaciones;
 
 import domain.heladera.Heladera;
+import domain.usuarios.Colaborador;
+import domain.usuarios.ColaboradorFisico;
 
 import java.time.LocalDate;
 
-public class Distribuir implements Donable {
+public class Distribuir  {
     private Heladera heladeraOrigen;
     private Heladera heladeraDestino;
     private Integer cantidad;
     private Motivo motivo;
     private LocalDate fechaDeDonacion;
+    private ColaboradorFisico colaboradorQueLaDono;
 
 
-    @Override
-    public boolean puedeserDonada() {
-        return true;
-    }
 
-    @Override
-    public String msgError() {
-        return "La distribucion no puede ser donada";
-    }
-
-    public Distribuir(Heladera heladeraOrigen,Heladera heladeraDestino, Integer cantidad, Motivo motivo,LocalDate fechaDeDonacion){
+    public Distribuir(Heladera heladeraOrigen, Heladera heladeraDestino, Integer cantidad, Motivo motivo, LocalDate fechaDeDonacion, ColaboradorFisico colaboradorQuelaDono){
         this.heladeraOrigen = heladeraOrigen;
         this.heladeraDestino = heladeraDestino;
         this.cantidad = cantidad;
         this.motivo = motivo;
         this.fechaDeDonacion = fechaDeDonacion;
+        this.colaboradorQueLaDono = colaboradorQuelaDono;
+
+
     }
