--- src/main/java/domain/donaciones/Donable.java
@@ -1,7 +1,0 @@
-package domain.donaciones;
-
-public interface Donable {
-    boolean puedeserDonada();
-    String msgError();
-
-}
