+++ src/main/java/domain/usuarios/ColaboradorFisico.java
@@ -1,56 +1,28 @@
 package domain.usuarios;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import domain.contacto.MedioDeContacto;
-import domain.donaciones.Donable;
-import domain.formulario.Formulario;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.Setter;
 import lombok.ToString;
 
 @ToString
-public class ColaboradorFisico {
+public class ColaboradorFisico extends Colaborador {
 
     @NonNull @Getter @Setter  private String nombre;
     @NonNull @Getter @Setter private String apellido;
-    @Getter @Setter private Formulario formulario;
-    @Getter private List <Donable> donacionesRealizadas = new ArrayList<>();
-
-
-    @Getter private List <MedioDeContacto> medioContacto; //value object, que al menos complete un valor
 
     public ColaboradorFisico(String nombre, String apellido) {
         this.nombre = nombre;
         this.apellido = apellido;
     }
 
-    /*Punto de arranque*/
-    public void completarFormulario(){
-        this.formulario = new Formulario();
-    }
 
-    /*Sobrecarga*/
-    public void completarFormulario(Formulario formulario){
-        this.formulario = formulario;
-    }
-
-    public void donar(Donable donacion) {
-        if (!donacion.puedeserDonada()){
-            throw new RuntimeException(donacion.msgError());
-        }
-        donacionesRealizadas.add(donacion);
-        //formulario.registrarDonacion(donacion);
-    }
-
-    public void agregarRespuesta(String label, String valor) {
-        formulario.cargarValor(label,valor);
-    }
+//    public void donar(Donable donacion) {
+//        if (!donacion.puedeserDonada()){
+//            throw new RuntimeException(donacion.msgError());
+//        }
+//        donacionesRealizadas.add(donacion);
+//    }
 
 
-    public void leerFormulario(){
-        formulario.mostrarCampos();
-    }
 }
